[Title] - Edit the Reflection
[Desc]
 - ** Fix ** the dead code elimination was occured in the gcc compiler on the template class that used the property and method macro.
Add the "used" macro keyword in the property and method's static variable that make the property info and method info.
But, this solution is not perfect, so, needs to find the right solution.
 - Add the Getter the property and method info from the super type.
So, if the properties and methods are same as super's things,
the hash(used the object's name) will be same and fail to add the type info.
So, Add the class name function in the utils file.
The file helps the type info get the real name and the property and method add the keyword that is the type name in front of the their name.
 - Fix the container property info's getter function for the mapped type. (m_keyType -> m_mappedType)
[Type/Branch]
[Auther/Data]